test success

+++ src/main/java/driving/model/DriveEvent.java
@@ -21,7 +21,7 @@ public class DriveEvent implements Serializable {
   @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
   LocalDateTime timestamp;
   MessageType messageType;
-  int messageSequenceNumber;
+  long messageSequenceNumber;
   List<DriveCoordinate> coordinates;
 
   public String transactionId() {
